Fix error handling, bump goMesh
- Fix error handling in certain code paths.
- Bump goMesh to 0.2.0 to include message stall fix
- Change location to use int values

+++ location.go
@@ -9,5 +9,5 @@ func setLocation(c *cli.Context) error {
 	defer radio.Close()
 
 	alt := int32(c.Int("alt"))
-	return radio.SetLocation(c.Float64("lat"), c.Float64("long"), alt)
+	return radio.SetLocation(int32(c.Int64("lat")), int32(c.Int64("long")), alt)
 }
